pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "versioningit~=2.0"]
build-backend = "setuptools.build_meta"

# Self-descriptive entries which should always be present
# https://packaging.python.org/en/latest/specifications/declaring-project-metadata/
[project]
name = "pympfit"
description = "Partial atomic charge assignment via multipole moment-based fitting algorithm"
dynamic = ["version"]
readme = "README.md"
authors = [
    { name = "Shehan M. Parmar", email = "sparmar32@gatech.edu" }
]
license = "MIT"
license-files = ["LICENSE"]
# See https://pypi.org/classifiers/
classifiers = [
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.10"
dependencies = [
    "openff-recharge>=0.5.0",
    "openff-utilities",
    "scipy>=1.7",
]

[project.urls]
"Source" = "https://github.com/shehan807/pympfit"
"Documentation" = "https://pympfit.readthedocs.io/"

[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-xdist>=3.0",
]

[tool.setuptools]
# This subkey is a beta stage development and keys may change in the future, see https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html for more details
#
# As of version 0.971, mypy does not support type checking of installed zipped
# packages (because it does not actually import the Python packages).
# We declare the package not-zip-safe so that our type hints are also available
# when checking client code that uses our (installed) package.
# Ref:
# https://mypy.readthedocs.io/en/stable/installed_packages.html?highlight=zip#using-installed-packages-with-mypy-pep-561
zip-safe = false
# Let setuptools discover the package in the current directory,
# but be explicit about non-Python files.
# See also:
#   https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html#setuptools-specific-configuration
# Note that behavior is currently evolving with respect to how to interpret the
# "data" and "tests" subdirectories. As of setuptools 63, both are automatically
# included if namespaces is true (default), even if the package is named explicitly
# (instead of using 'find'). With 'find', the 'tests' subpackage is discovered
# recursively because of its __init__.py file, but the data subdirectory is excluded
# with include-package-data = false and namespaces = false.
include-package-data = false
[tool.setuptools.packages.find]
namespaces = false
where = ["."]

# Ref https://setuptools.pypa.io/en/latest/userguide/datafiles.html#package-data
[tool.setuptools.package-data]
pympfit = [
    "py.typed",
    "data/**/*",
]

[tool.versioningit]
default-version = "1+unknown"

[tool.versioningit.format]
distance = "{base_version}+{distance}.{vcs}{rev}"
dirty = "{base_version}+{distance}.{vcs}{rev}.dirty"
distance-dirty = "{base_version}+{distance}.{vcs}{rev}.dirty"

[tool.versioningit.vcs]
# The method key:
method = "git"  # <- The method name
# Parameters to pass to the method:
match = ["*"]
default-tag = "1.0.0"

[tool.versioningit.write]
file = "pympfit/_version.py"

[tool.ruff]
target-version = "py310"
output-format = "concise"
line-length = 88
exclude = ["pympfit/_version.py"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "ANN002",  # Missing type annotation for *arg
    "ANN003",  # Missing type annotation for **kwargs
    "ANN401",  # typing.Any disallowed
    "ARG002",  # unused method argument
    "C408",    # Unnecessary (dict/list/tuple) call - remove call
    "C901",    # function too complex
    "COM812",  # trailing comma missing
    "EM",      # exception message must not use f-string literal
    "ERA001",  # found commented out code
    "FBT001",  # boolean positional arg in function definition
    "FBT002",  # boolean default value in function definition
    "FIX002",  # line contains TODO
    "G004",    # logging uses f-string
    "ISC001",  # implicit string concatenation
    "PD011",   # pandas-use-of-dot-values
    "PERF203", # try-except-in-loop
    "PLC0415", # imports not at top level are OK
    "PLR0911", # too many returns
    "PLR0912", # too many branches
    "PLR0913", # too many arguments
    "PLR0915", # too many local statements
    "PLR2004", # magic value comparison
    "PT006",   # pytest-parametrize-names-wrong-type
    "PT013",   # pytest-incorrect-pytest-import
    "PTH",     # prefer Pathlib to os.path
    "RUF013",  # implicit-optional
    "S324",    # use of insecure hash function
    "S507",    # paramiko auto trust
    "S603",    # use of insecure subprocess
    "TD",      # todos
    "TRY003",  # long message outside exception class
    "SLF001",  # private member access (using internal _classes is fine in own codebase)
]
pydocstyle.convention = "numpy"
isort.known-first-party = ["pympfit"]
isort.split-on-trailing-comma = false

[tool.ruff.format]
docstring-code-format = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"**/tests/*" = ["ANN", "ARG001", "D", "E741", "INP001", "PLR2004", "S101"]
"docs/*" = ["INP001", "D100", "A001"]
# Devtools: MolSSI cookiecutter template scripts
"devtools/*" = ["ANN", "B904", "D", "E501", "INP001", "T201", "S602"]
# Examples: demonstration scripts with intentional print
"examples/*" = ["ANN", "D", "INP001", "T201"]
# Legacy files: l is physics convention for angular momentum, W is standard notation, print is intentional
"pympfit/mpfit/_legacy.py" = ["E741", "T201"]
"pympfit/mpfit/_legacy_constrained.py" = ["E741", "N806", "T201"]
# Core math: l/m are quantum numbers, A/Y/W are standard physics notation, m used for loop indexing
"pympfit/mpfit/core.py" = ["B007", "E741", "N802", "N803", "N806"]
# Solvers: U/S/Vh are SVD notation, TRY301 raises in try are intentional
"pympfit/mpfit/solvers.py" = ["E741", "N806", "TRY301"]
# Optimize: A_matrix/site_A/site_b are standard physics matrix notation
"pympfit/optimize/_optimize.py" = ["N806"]
# Psi4: S607 partial executable path is intentional (psi4 expected in PATH)
"pympfit/gdma/psi4.py" = ["S607"]

[tool.coverage.run]
omit = [
    "*/tests/*",
    "pympfit/_version.py",
]

[tool.coverage.report]
exclude_lines = [
    "@overload",
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.mypy]
plugins = ["numpy.typing.mypy_plugin"]
warn_unused_configs = true
warn_unused_ignores = true
show_error_codes = true
disable_error_code = ["call-arg"]
exclude = ["pympfit/tests/"]

[tool.pytest.ini_options]
testpaths = ["pympfit/tests"]
addopts = "-v --tb=short -m 'not slow'"
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
